fix: use createResizedImage function to decrease download avatar image

+++ src/hooks/media/preview.ts
@@ -12,9 +12,9 @@ import {
   readFile,
 } from '@2060/utils/RNFS'
 
-const LOCAL_PREVIEW_IMAGE_WIDTH = 512
+export const LOCAL_PREVIEW_IMAGE_WIDTH = 512
 const LOCAL_PREVIEW_IMAGE_HEIGHT = 512
-const LOCAL_PREVIEW_IMAGE_QUALITY = 70
+export const LOCAL_PREVIEW_IMAGE_QUALITY = 70
 
 const DIDCOMM_PREVIEW_IMAGE_WIDTH = 128
 const DIDCOMM_PREVIEW_IMAGE_HEIGHT = 128
@@ -32,17 +32,17 @@ export async function createLocalPreview(options: { mimeType: string; localFileP
     })
     localPreviewFilePath = thumbnailResponse?.path
   } else if (mimeType.startsWith('image')) {
-    const previewResponse = await createImagePreview({
+    const previewResponse = await createResizedImage({
       imageUrl: localFilePath,
       maxWidth: LOCAL_PREVIEW_IMAGE_WIDTH,
       maxHeight: LOCAL_PREVIEW_IMAGE_HEIGHT,
       quality: LOCAL_PREVIEW_IMAGE_QUALITY,
     })
-    localPreviewFilePath = previewResponse.path
+    if (previewResponse) localPreviewFilePath = previewResponse.path
   }
 
   // save preview under previews directory
-  // TODO: this should be actually done directly by createVideoThumbnail/createImagePreview
+  // TODO: this should be actually done directly by createVideoThumbnail/createResizedImage
   if (localPreviewFilePath) {
     const [previewFileName] = localFilePath.split('/').slice(-1)
     const previewMediaDestinationPath = `${getLocalMediaPreviewFilePath(previewFileName)}.jpeg`
@@ -76,16 +76,16 @@ export async function createDidCommPreview(options: { mimeType: string; localFil
       await deleteFile(thumbnailResponse.path)
     }
   } else if (mimeType.startsWith('image')) {
-    const previewResponse = await createImagePreview({
+    const previewResponse = await createResizedImage({
       imageUrl: localFilePath,
       maxWidth: DIDCOMM_PREVIEW_IMAGE_WIDTH,
       maxHeight: DIDCOMM_PREVIEW_IMAGE_HEIGHT,
       quality: DIDCOMM_PREVIEW_IMAGE_QUALITY,
     })
-    didcommPreview = previewResponse.base64
-
-    // Clean up file
-    await deleteFile(previewResponse.path)
+    if (previewResponse) {
+      didcommPreview = previewResponse.base64
+      await deleteFile(previewResponse.path)
+    }
   }
 
   return didcommPreview
@@ -113,21 +113,25 @@ async function createVideoThumbnail(options: {
   }
 }
 
-async function createImagePreview(options: {
+export async function createResizedImage(options: {
   imageUrl: string
   maxWidth?: number
   maxHeight?: number
   quality?: number
 }) {
-  const { imageUrl, maxWidth, maxHeight, quality } = options
-  const preview = await ImageResizer.createResizedImage(
-    imageUrl,
-    maxWidth ?? DIDCOMM_PREVIEW_IMAGE_WIDTH,
-    maxHeight ?? DIDCOMM_PREVIEW_IMAGE_HEIGHT,
-    'JPEG',
-    quality ?? DIDCOMM_PREVIEW_IMAGE_QUALITY,
-  )
-
-  const data = await readFile(preview.path, 'base64')
-  return { path: preview.path, base64: dataUrl('image/jpeg', data) }
+  try {
+    const { imageUrl, maxWidth, maxHeight, quality } = options
+    const preview = await ImageResizer.createResizedImage(
+      imageUrl,
+      maxWidth ?? DIDCOMM_PREVIEW_IMAGE_WIDTH,
+      maxHeight ?? DIDCOMM_PREVIEW_IMAGE_HEIGHT,
+      'JPEG',
+      quality ?? DIDCOMM_PREVIEW_IMAGE_QUALITY,
+    )
+    const data = await readFile(preview.path, 'base64')
+    return { path: preview.path, base64: dataUrl('image/jpeg', data) }
+  } catch (error) {
+    logError(`Error creating image preview: ${error}`)
+    return null
+  }
 }
